backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow>=10.3.0
openai>=1.3.5
python-dotenv==1.0.0
werkzeug==3.0.1
